+++ src/main/java/com/rafaelhayashi/brasilprev/controller/CustomerController.java
@@ -1,12 +1,16 @@
 package com.rafaelhayashi.brasilprev.controller;
 
+import com.rafaelhayashi.brasilprev.controller.dto.CustomerDetailsDto;
+import com.rafaelhayashi.brasilprev.controller.dto.CustomerDto;
+import com.rafaelhayashi.brasilprev.controller.form.CustomerAddressForm;
+import com.rafaelhayashi.brasilprev.controller.form.CustomerForm;
+import com.rafaelhayashi.brasilprev.controller.form.CustomerUpdateForm;
 import com.rafaelhayashi.brasilprev.model.Customer;
 import com.rafaelhayashi.brasilprev.service.CustomerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Repository;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
