+++ src/main/java/com/rafaelhayashi/brasilprev/controller/dto/CustomerDto.java
@@ -1,4 +1,4 @@
-package com.rafaelhayashi.brasilprev.controller;
+package com.rafaelhayashi.brasilprev.controller.dto;
 
 import com.rafaelhayashi.brasilprev.model.Customer;
 import org.springframework.data.domain.Page;
