+++ src/main/java/com/rafaelhayashi/brasilprev/controller/form/CustomerUpdateForm.java
@@ -1,4 +1,4 @@
-package com.rafaelhayashi.brasilprev.controller;
+package com.rafaelhayashi.brasilprev.controller.form;
 
 import javax.validation.constraints.NotEmpty;
